requirements.txt
minio
apache-airflow-providers-amazon